+++ scripts/embedding/gprolog/gprolog_embed_logtalk.sh
@@ -5,7 +5,7 @@
 ##   This script creates a new GNU Prolog top-level
 ##   interpreter that embeds Logtalk
 ## 
-##   Last updated on April 2, 2018
+##   Last updated on April 3, 2018
 ## 
 ##   This file is part of Logtalk <https://logtalk.org/>  
 ##   Copyright 1998-2018 Paulo Moura <pmoura@logtalk.org>
@@ -65,33 +65,41 @@ elif ! [ -d "$LOGTALKHOME" ]; then
 fi
 
 print_version() {
-	echo "$(basename "$0") 0.1"
+	echo "$(basename "$0") 0.3"
 	exit 0
 }
 
 usage_help()
 {
 	echo 
-	echo "This script creates a new GNU Prolog top-level interpreter that embeds Logtalk"
+	echo "This script creates a new GNU Prolog top-level interpreter that embeds the"
+	echo "Logtalk compiler and runtime and an optional application from an application"
+	echo "source code given its loader file."
 	echo
 	echo "Usage:"
 	echo "  $(basename "$0") [-d directory]"
+	echo "  $(basename "$0") [-l loader]"
+	echo "  $(basename "$0") [-s settings]"
 	echo "  $(basename "$0") -v"
 	echo "  $(basename "$0") -h"
 	echo
 	echo "Optional arguments:"
 	echo "  -v print version of $(basename "$0")"
 	echo "  -d directory to use for intermediate and final results (default is $HOME/collect)"
+	echo "  -l optional loader file for the application"
+	echo "  -s optional settings file for the application"
 	echo "  -h help"
 	echo
 	exit 0
 }
 
-while getopts "vd:h" option
+while getopts "vd:l:s:h" option
 do
 	case $option in
 		v) print_version;;
 		d) d_arg="$OPTARG";;
+		l) l_arg="$OPTARG";;
+		s) s_arg="$OPTARG";;
 		h) usage_help;;
 		*) usage_help;;
 	esac
@@ -99,6 +107,28 @@ done
 
 if [ "$d_arg" != "" ] ; then
 	directory="$d_arg"
+fi
+
+if [ "$l_arg" != "" ] ; then
+	if [ -f "$l_arg" ] ; then
+		loader="$l_arg"
+	else
+		echo "The $l_arg loader file does not exist!"
+		exit 1
+	fi
+else
+	loader=""
+fi
+
+if [ "$s_arg" != "" ] ; then
+	if [ -f "$s_arg" ] ; then
+		settings="$s_arg"
+	else
+		echo "The $s_arg settings file does not exist!"
+		exit 1
+	fi
+else
+	settings=""
 fi
 
 mkdir -p "$directory"
@@ -122,6 +152,23 @@ cp "$LOGTALKHOME/adapters/gnu.pl" .
 cp "$LOGTALKHOME/paths/paths_core.pl" .
 cp "$LOGTALKHOME/core/core.pl" .
 
-gplc -o logtalk gnu.pl expanding*_lgt.pl monitoring*_lgt.pl forwarding*_lgt.pl user*_lgt.pl logtalk*_lgt.pl core_messages*_lgt.pl core.pl paths_core.pl
+if [ "$settings" != "" ] ; then
+	gplgt$extension --query-goal "logtalk_compile('$settings',[optimize(on),scratch_directory('$directory')]),halt"
+else
+	touch settings_lgt.pl
+fi
+
+if [ "$loader" != "" ] ; then
+	mkdir -p "$directory/application"
+	cd "$directory/application"
+	gplgt$extension --query-goal "set_logtalk_flag(source_data, off),set_logtalk_flag(optimize, on),set_logtalk_flag(clean, off),set_logtalk_flag(context_switching_calls, deny),set_logtalk_flag(scratch_directory, '$directory/application'),logtalk_load('$loader', [clean(on)]),halt"
+	rm -f $directory/application/*loader*
+	cd ..
+else
+	touch application.pl
+fi
+
+gplc -o logtalk gnu.pl expanding*_lgt.pl monitoring*_lgt.pl forwarding*_lgt.pl user*_lgt.pl logtalk*_lgt.pl core_messages*_lgt.pl settings*_lgt.pl core.pl paths_core.pl $(ls application/*.pl)
 
 rm *.pl
+rm application/*.pl